requirements.txt
aiofiles==24.1.0
fastapi==0.115.5
Jinja2==3.1.4
openai==1.55.3
python-dotenv==1.0.1
python-multipart==0.0.18
uvicorn==0.32.1
websockets==14.1
